Worked on cross-compiler-stuff. shlibtest works again :)

+++ lib/cpp/src/runtime/new.cpp
@@ -20,11 +20,11 @@
 #include <esc/common.h>
 #include <stdlib.h>
 
-void *operator new(u32 size) {
+void *operator new(size_t size) {
     return malloc(size);
 }
 
-void *operator new[](u32 size) {
+void *operator new[](size_t size) {
     return malloc(size);
 }
